Code cleanup and improvements
